requirements.txt
psycopg2-binary
Faker
python-dotenv

